requirement.txt

# pip install -r requirement.txt
Django==5.2
pillow==11.1.0
djangorestframework==3.16.0